Renames module and tests
